Add new ExpressionOptions.StringConcat to concat values as string
Add new `ExpressionOptions.StringConcat` option to concat values as string.
This is a breaking change. In previous versions if the left value is string the result was a string concatenation, but if the left is not string the result will be an arithmetic addition. After this change you should explicitly specify this option for string concatenation.